z*range commands now receive long instead of int. #242

+++ src/main/java/redis/clients/jedis/ShardedJedis.java
@@ -298,7 +298,7 @@ public class ShardedJedis extends BinaryShardedJedis implements JedisCommands {
         return j.zadd(key, score, member);
     }
 
-    public Set<String> zrange(String key, int start, int end) {
+    public Set<String> zrange(String key, long start, long end) {
         Jedis j = getShard(key);
         return j.zrange(key, start, end);
     }
@@ -323,17 +323,17 @@ public class ShardedJedis extends BinaryShardedJedis implements JedisCommands {
         return j.zrevrank(key, member);
     }
 
-    public Set<String> zrevrange(String key, int start, int end) {
+    public Set<String> zrevrange(String key, long start, long end) {
         Jedis j = getShard(key);
         return j.zrevrange(key, start, end);
     }
 
-    public Set<Tuple> zrangeWithScores(String key, int start, int end) {
+    public Set<Tuple> zrangeWithScores(String key, long start, long end) {
         Jedis j = getShard(key);
         return j.zrangeWithScores(key, start, end);
     }
 
-    public Set<Tuple> zrevrangeWithScores(String key, int start, int end) {
+    public Set<Tuple> zrevrangeWithScores(String key, long start, long end) {
         Jedis j = getShard(key);
         return j.zrevrangeWithScores(key, start, end);
     }
@@ -408,7 +408,7 @@ public class ShardedJedis extends BinaryShardedJedis implements JedisCommands {
         return j.zrevrangeByScoreWithScores(key, max, min, offset, count);
     }
 
-    public Long zremrangeByRank(String key, int start, int end) {
+    public Long zremrangeByRank(String key, long start, long end) {
         Jedis j = getShard(key);
         return j.zremrangeByRank(key, start, end);
     }
